finished fixing
