+++ .github/workflows/deploy.yml
@@ -28,10 +28,7 @@ jobs:
       - name: Set up Node.js
         uses: actions/setup-node@v4
         with:
-          node-version: '20.x'
-
-      - name: Install dependencies
-        run: npm install
+          node-version: '18.x'
 
       - name: Set up Python
         uses: actions/setup-python@v2
