+++ test.js
@@ -1,4 +1,4 @@
-import { extractJobs } from './functions/scraper.js';
+import { extractJobs } from './functions/scraper.mjs';
 
 extractJobs().then(jobs => console.log(jobs));
